FeatureLocation/PageHeader are not Writer specific
relocate their definitions to a more neutral location
